+++ src/features/upload/slice.ts
@@ -7,14 +7,15 @@ import { type Action, createSlice, type PayloadAction } from '@reduxjs/toolkit';
 
 import { SHOW_UPLOAD_JOB } from '~/features/show/constants';
 import { _clearLoadedShow } from '~/features/show/slice';
-import type UAV from '~/model/uav';
-import { type Collection, replaceItemOrAddToFront } from '~/utils/collections';
+import type { Identifier } from '~/utils/collections';
 import { noPayload } from '~/utils/redux';
 
-import type { JobData, JobPayload, UploadJob } from './types';
+import type { HistoryItem, JobData, JobPayload } from './types';
 import {
-  clearLastUploadResultForJobTypeHelper,
   clearQueues,
+  clearUploadHistoryForJobTypeHelper,
+  compactHistory,
+  createHistoryItem,
   ensureItemsInQueue,
   moveItemsBetweenQueues,
 } from './utils';
@@ -37,34 +38,28 @@ export type UploadSliceState = {
   };
 
   /**
-   * History of recent upload jobs. Each upload job has a _type_ and a
-   * _hash_; the type identifies the type of the job (e.g., show upload,
-   * parameter upload etc) while the key is a compact representation of the
-   * exact data that was uploaded such that different jobs of the same type
-   * have different keys. In the absence of such a hash, use a sequential
-   * counter. The ID of the job should be the type and we only keep the
-   * latest job from each type in the history.
+   * Record that maps job types to the corresponding upload history items.
    */
-  history: Collection<UploadJob>;
+  history: Record<string, HistoryItem[]>;
 
   queues: {
-    itemsInProgress: Array<UAV['id']>;
-    itemsWaitingToStart: Array<UAV['id']>;
-    itemsQueued: Array<UAV['id']>;
-    itemsFinished: Array<UAV['id']>;
-    failedItems: Array<UAV['id']>;
+    itemsInProgress: Identifier[];
+    itemsWaitingToStart: Identifier[];
+    itemsQueued: Identifier[];
+    itemsFinished: Identifier[];
+    failedItems: Identifier[];
   };
 
   // If you add a new queue above, make sure that the ALL_QUEUES array
   // is updated in features/upload/utils.js
 
   /** Errors corresponding to the individual UAVs in the current upload job */
-  errors: Record<UAV['id'], unknown>;
+  errors: Record<Identifier, string>;
 
   /** Progress information corresponding to the individual UAVs in the
    * current upload job.
    */
-  progresses: Record<UAV['id'], number>;
+  progresses: Record<Identifier, number>;
 
   /**
    * Timing information related to the current upload job, required to
@@ -109,10 +104,7 @@ const initialState: UploadSliceState = {
     running: false,
   },
 
-  history: {
-    order: [],
-    byId: {},
-  },
+  history: {},
   queues: {
     itemsInProgress: [],
     itemsWaitingToStart: [],
@@ -146,10 +138,17 @@ const { actions, reducer } = createSlice({
   name: 'upload',
   initialState,
   reducers: {
-    clearLastUploadResultForJobType(state, action: PayloadAction<string>) {
+    clearUploadHistoryForCurrentJobType(state) {
+      const jobType = state.dialog.selectedJob.type;
+      if (jobType) {
+        clearUploadHistoryForJobTypeHelper(state, jobType);
+      }
+    },
+
+    clearUploadHistoryForJobType(state, action: PayloadAction<string>) {
       const { payload: jobType } = action;
       if (jobType) {
-        clearLastUploadResultForJobTypeHelper(state, jobType);
+        clearUploadHistoryForJobTypeHelper(state, jobType);
       }
     },
 
@@ -180,7 +179,7 @@ const { actions, reducer } = createSlice({
       action: PayloadAction<{
         type: string;
         payload: JobPayload;
-        targets: Array<UAV['id']>;
+        targets: Identifier[];
       }>
     ) {
       const { payload } = action;
@@ -194,7 +193,7 @@ const { actions, reducer } = createSlice({
       state.currentJob.type = type;
       state.currentJob.payload = jobPayload;
 
-      clearQueues(state);
+      clearQueues(state, { showLastUploadResult: false });
 
       state.queues.itemsWaitingToStart = [...targets];
     },
@@ -224,29 +223,31 @@ const { actions, reducer } = createSlice({
       action: PayloadAction<{ success: boolean; cancelled: boolean }>
     ) {
       const { success, cancelled } = action.payload;
+      const jobType = state.currentJob.type;
 
       // Dispatched by the saga; should not be dispatched manually
 
-      // Clear the queues
-      state.queues.itemsWaitingToStart = [];
-      state.queues.itemsInProgress = [];
-      state.queues.itemsQueued = [];
+      // Store the data in history
+      if (jobType) {
+        const historyItem = createHistoryItem(
+          cancelled ? 'cancelled' : success ? 'success' : 'error',
+          state.queues,
+          state.errors
+        );
+
+        if (!state.history[jobType]) {
+          state.history[jobType] = [];
+        }
+        state.history[jobType].push(historyItem);
+        state.history[jobType] = compactHistory(state.history[jobType]);
+      }
+
+      // Clear queues and show the last upload result in the dialog.
+      // Everything related to the job is persisted in the history.
+      clearQueues(state, { showLastUploadResult: true });
 
       // Reset the current job to an idle state
       state.currentJob.running = false;
-
-      // Store the upload result in the history
-      if (state.currentJob.type) {
-        const historyItem: UploadJob = {
-          id: state.currentJob.type,
-          payload: state.currentJob.payload,
-          result: cancelled ? 'cancelled' : success ? 'success' : 'error',
-        };
-        replaceItemOrAddToFront(state.history, historyItem);
-      }
-
-      // Trigger the dialog box to show the result
-      state.dialog.showLastUploadResult = true;
     },
 
     _notifyUploadStartedAt(state, action: PayloadAction<number>) {
@@ -288,7 +289,7 @@ const { actions, reducer } = createSlice({
     _setErrorMessageForUAV: {
       reducer(
         state,
-        action: PayloadAction<{ uavId: UAV['id']; message: string }>
+        action: PayloadAction<{ uavId: Identifier; message: string }>
       ) {
         const { uavId, message } = action.payload;
         if (message) {
@@ -298,7 +299,7 @@ const { actions, reducer } = createSlice({
         }
       },
 
-      prepare: (uavId: UAV['id'], message: string | Error) => ({
+      prepare: (uavId: Identifier, message: string | Error) => ({
         payload: { uavId, message: String(message) },
       }),
     },
@@ -315,7 +316,7 @@ const { actions, reducer } = createSlice({
     _setProgressInfoForUAV: {
       reducer(
         state,
-        action: PayloadAction<{ uavId: UAV['id']; progress: number }>
+        action: PayloadAction<{ uavId: Identifier; progress: number }>
       ) {
         const { uavId, progress } = action.payload;
         if (progress >= 0 && progress <= 1) {
@@ -325,7 +326,7 @@ const { actions, reducer } = createSlice({
         }
       },
 
-      prepare: (uavId: UAV['id'], progress: number) => ({
+      prepare: (uavId: Identifier, progress: number) => ({
         payload: { uavId, progress },
       }),
     },
@@ -368,7 +369,7 @@ const { actions, reducer } = createSlice({
         !state.currentJob.running &&
         state.dialog.selectedJob?.type !== newJobType
       ) {
-        clearQueues(state);
+        clearQueues(state, { showLastUploadResult: false });
       }
 
       state.dialog.backAction = backAction;
@@ -407,14 +408,15 @@ const { actions, reducer } = createSlice({
 
   extraReducers(builder) {
     builder.addCase(_clearLoadedShow, (state) => {
-      clearLastUploadResultForJobTypeHelper(state, SHOW_UPLOAD_JOB.type);
+      clearUploadHistoryForJobTypeHelper(state, SHOW_UPLOAD_JOB.type);
     });
   },
 });
 
 export const {
   cancelUpload,
-  clearLastUploadResultForJobType,
+  clearUploadHistoryForCurrentJobType,
+  clearUploadHistoryForJobType,
   clearUploadQueue,
   closeUploadDialog,
   dismissLastUploadResult,