+++ src/features/upload/types.ts
@@ -15,8 +15,19 @@ export type JobData = {
   payload?: JobPayload;
 };
 
-export type UploadJob = {
-  id: Identifier;
-  payload: JobPayload;
-  result: 'success' | 'error' | 'cancelled';
+export type UAVStatus = 'success' | 'error';
+
+export type UploadJobResult = UAVStatus | 'cancelled';
+
+/**
+ * Aggregated upload status for a job type.
+ */
+export type UploadStatus = UAVStatus | 'partial' | 'not-available';
+
+type ErrorMessage = string;
+
+export type HistoryItem = {
+  result: UploadJobResult;
+  perUavStatuses: Record<Identifier, UAVStatus>;
+  perUavErrors: Record<Identifier, ErrorMessage>;
 };