+++ src/views/show-control/ShowUploadDialogButton.tsx
@@ -1,5 +1,6 @@
-import ListItemButton from '@mui/material/ListItemButton';
-import PropTypes from 'prop-types';
+import ListItemButton, {
+  type ListItemButtonProps,
+} from '@mui/material/ListItemButton';
 import { useTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
 
@@ -7,20 +8,26 @@ import { StatusLight } from '@skybrush/mui-components';
 
 import ListItemTextWithProgress from '~/components/ListItemTextWithProgress';
 import { Status } from '~/components/semantics';
-import { SHOW_UPLOAD_JOB } from '~/features/show/constants';
+import { JOB_TYPE, SHOW_UPLOAD_JOB } from '~/features/show/constants';
 import { getSetupStageStatuses } from '~/features/show/stages';
 import {
-  getUploadProgress,
+  getSelectedJobTypeInUploadDialog,
   isUploadInProgress,
 } from '~/features/upload/selectors';
 import { openUploadDialogForJob } from '~/features/upload/slice';
 import UploadProgressBar from '~/features/upload/UploadProgressBar';
+import type { RootState } from '~/store/reducers';
+
+type Props = Omit<ListItemButtonProps, 'disabled'> & {
+  loading: boolean;
+  status: Status;
+};
 
 /**
  * React component for the button that allows the user to start or stop the
  * upload process of the current show to the drones.
  */
-const ShowUploadDialogButton = ({ loading, status, ...rest }) => {
+const ShowUploadDialogButton = ({ loading, status, ...rest }: Props) => {
   const { t } = useTranslation();
 
   return (
@@ -38,20 +45,13 @@ const ShowUploadDialogButton = ({ loading, status, ...rest }) => {
   );
 };
 
-ShowUploadDialogButton.propTypes = {
-  loading: PropTypes.bool,
-  onClick: PropTypes.func,
-  status: PropTypes.oneOf(Object.values(Status)),
-};
-
 export default connect(
-  // mapStateToProps
-  (state) => ({
-    loading: isUploadInProgress(state),
-    progress: getUploadProgress(state),
+  (state: RootState) => ({
+    loading:
+      getSelectedJobTypeInUploadDialog(state) === JOB_TYPE &&
+      isUploadInProgress(state),
     status: getSetupStageStatuses(state).uploadShow,
   }),
-  // mapDispatchToProps
   {
     onClick: () => openUploadDialogForJob({ job: SHOW_UPLOAD_JOB }),
   }